Changes for running with docker

+++ lib/tasks/load_fixtures.rake
@@ -2,17 +2,20 @@ desc 'Load the fixtures into the database'
 task :load_fixtures, [] => :environment do |t, args|
 	require 'csv'
 	EPSILON = 0.00000001
+	
 	# Load all the models
 	Dir[Rails.root.join("app", "models", "**", "*.rb")].each(&method(:require))
-	# models = ActiveRecord::Base.subclasses.index_by(&:table_name)
 	models = ::ApplicationRecord.subclasses.index_by(&:table_name)
-	puts models
+	
 	Dir[Rails.root.join('db', 'fixtures', '*.csv')].each do |file|
-		puts file
 		table_name = file.split('/').last.split('.').first
-		puts table_name
 		model = models[table_name]
-		puts model
+		puts "Loading #{model}..."
+		
+		if model.all.length > 5
+			puts "#{model} already loaded"
+			next
+		end
 
 		# Clear out existing records
 		model.delete_all
@@ -27,13 +30,14 @@ task :load_fixtures, [] => :environment do |t, args|
 				header = row
 				next
 			end
+			
 			if ( (i / line_count * 100.0 + 1.0) % 5 ) - 1.0 < Float::EPSILON
 				puts "#{(i / line_count * 100.0).round}% done..."
-			else
-				# puts "#{( (i / line_count * 100.0 + 1.0) % 5 ) - 1.0}"
 			end
+			
 			atts = {}
 			header.each_with_index{|h,i| atts[h.to_sym] = row[i]}
+			
 			begin
 				model.create(atts)
 			rescue Exception => err
@@ -41,6 +45,7 @@ task :load_fixtures, [] => :environment do |t, args|
 				puts "\n Model data:\n#{atts}"
 				puts "\n Exception:\n#{err.msg}"
 			end
+			
 			i += 1.0
 		end
 		puts "#{table_name} successfully seeded."
